requirements.txt
# 吹き出し抽出手法評価システム - 必要パッケージ

# 基本パッケージ
numpy>=1.20.0
opencv-python>=4.5.0
scipy>=1.7.0

# 画像処理・機械学習
scikit-image>=0.18.0
scikit-learn>=0.24.0

# 可視化
matplotlib>=3.3.0

# データ処理
Pillow>=8.0.0

# オプション: より高速な計算（推奨）
# numba>=0.53.0
